+++ src/db/queryHelpers/order.js
@@ -35,7 +35,7 @@ export const updateOrder = async (id, update) => {
   }
 };
 
-// Get car
+// Get orders
 export const getOrder = async (condition) => {
   const value = Object.values(condition);
   const key = Object.keys(condition).toString();
@@ -45,7 +45,15 @@ export const getOrder = async (condition) => {
   };
   try {
     const { rows } = await pool.query(query);
-    return { error: null, result: rows };
+    // postgresSQL returns all entity as string, converting number values to number
+    const final = rows.map(row => ({
+      ...row,
+      amount: +row.amount,
+      car_price: +row.car_price,
+      buyer: +row.buyer,
+      car_id: +row.car_id,
+    }));
+    return { error: null, result: final };
   } catch (error) {
     return { error: error.message };
   }
