+++ src/routes/api/order.js
@@ -8,7 +8,7 @@ import { addOrder, updateOrder, getOrder } from '../../db/queryHelpers/order';
 const router = express.Router();
 
 // @route POST /api/v1/order
-// @desc Create an order request
+// @desc Create an order
 // @access Private, only authenticated users can make orders
 router.post('/', authMiddleware, async (req, res) => {
   const { errors, isValid } = carOrderValidator(req.body);
@@ -33,23 +33,28 @@ router.post('/', authMiddleware, async (req, res) => {
       if (car.owner === req.userData.id) {
         return res.status(400).json({ status: 400, error: 'Cannot place order for your own advert' });
       }
-      newOrder.car_price = car.price;
-      const { error, result } = await addOrder(newOrder);
-      if (!error) {
-        // format response to numeric types
-        const resData = {
-          id: result.id,
-          car_id: +result.car_id,
-          created_on: result.created_on,
-          status: result.status,
-          price: +result.car_price,
-          price_offered: +result.amount,
-        };
-        return res.status(201).json({ status: 201, data: resData });
+      // before adding a new order check if order for the same car already exist
+      const { result: privOrder } = await getOrder({ car_id: car.id });
+      if (!privOrder[0]) {
+        newOrder.car_price = car.price;
+        const { error, result } = await addOrder(newOrder);
+        if (!error) {
+          // format response to numeric types
+          const resData = {
+            id: result.id,
+            car_id: +result.car_id,
+            created_on: result.created_on,
+            status: result.status,
+            price: +result.car_price,
+            price_offered: +result.amount,
+          };
+          return res.status(201).json({ status: 201, data: resData });
+        }
+        return res.status(500).json({ status: 500, error: 'Server error' });
       }
-      return res.status(500).json({ status: 500, error: 'Server error' });
+      return res.status(409).json({ status: 409, error: `Order for car with id ${car.id} already exist` });
     }
-    return res.status(404).json({ status: 404, error: 'Car has already been sold' });
+    return res.status(404).json({ status: 404, error: 'Car not available' });
   }
   return res.status(404).json({ status: 404, error: 'Car with the given car_id does not exist' });
 });
@@ -94,5 +99,24 @@ router.patch('/:order_id/price', authMiddleware, async (req, res) => {
   }
   return res.status(400).json({ status: 400, error: 'parameter not understood' });
 });
+
+// @route GET /api/v1/order
+// @desc get users orders
+// @access Private, only authenticated users can have orders
+router.get('/', authMiddleware, async (req, res) => {
+  const { id: userId } = req.userData;
+  if (userId) {
+    const { result } = await getOrder({ buyer: userId });
+    if (result) {
+      return res.status(200).json({ status: 200, data: result });
+    }
+    // having userId and no result even an empty array, this must be a server error
+    return res.status(500).json({ status: 500, error: 'Server error' });
+  }
+  // how did this user even get through our middleware
+  return res.status(401).json({ status: 401, error: 'UnAuthorized user' });
+});
+
+// @todo -> users(buyer) should not place order for the same car twice
 
 export default router;
